Rename test-util modules to testutil
And make the package prefix match the module name.
And clean up the POMs a bit.
