+++ imagej/imagej-ops2/src/test/java/net/imagej/ops2/filter/hessian/HessianFilterTest.java
@@ -31,7 +31,7 @@ package net.imagej.ops2.filter.hessian;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import net.imagej.ops2.AbstractOpTest;
-import net.imagej.test_util.TestImgGeneration;
+import net.imagej.testutil.TestImgGeneration;
 import net.imglib2.Cursor;
 import net.imglib2.img.Img;
 import net.imglib2.type.numeric.real.FloatType;
