+++ imagej/imagej-ops2/src/test/java/net/imagej/ops2/image/integral/SquareIntegralImgTest.java
@@ -30,7 +30,7 @@
 package net.imagej.ops2.image.integral;
 
 import net.imagej.ops2.AbstractOpTest;
-import net.imagej.test_util.TestImgGeneration;
+import net.imagej.testutil.TestImgGeneration;
 import net.imglib2.RandomAccessibleInterval;
 import net.imglib2.img.Img;
 import net.imglib2.img.array.ArrayImg;
