+++ imagej/imagej-ops2/src/test/java/net/imagej/ops2/image/normalize/NormalizeTest.java
@@ -32,7 +32,7 @@ package net.imagej.ops2.image.normalize;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import net.imagej.ops2.AbstractOpTest;
-import net.imagej.test_util.TestImgGeneration;
+import net.imagej.testutil.TestImgGeneration;
 import net.imglib2.Cursor;
 import net.imglib2.IterableInterval;
 import net.imglib2.img.Img;
