+++ imagej/imagej-ops2/src/test/java/net/imagej/ops2/project/ProjectTest.java
@@ -32,7 +32,7 @@ package net.imagej.ops2.project;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import net.imagej.ops2.AbstractOpTest;
-import net.imagej.test_util.TestImgGeneration;
+import net.imagej.testutil.TestImgGeneration;
 import net.imglib2.RandomAccess;
 import net.imglib2.img.Img;
 import net.imglib2.type.numeric.integer.UnsignedByteType;
