+++ imagej/imagej-ops2/src/test/java/net/imagej/ops2/stats/DefaultMedianTest.java
@@ -33,7 +33,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 import net.imagej.ops2.AbstractOpTest;
-import net.imagej.test_util.TestImgGeneration;
+import net.imagej.testutil.TestImgGeneration;
 import net.imglib2.img.Img;
 import net.imglib2.type.numeric.RealType;
 import net.imglib2.type.numeric.integer.UnsignedByteType;
