+++ imagej/imagej-ops2/src/test/java/net/imagej/ops2/stats/StatisticsTest.java
@@ -30,7 +30,7 @@
 package net.imagej.ops2.stats;
 
 import net.imagej.ops2.AbstractOpTest;
-import net.imagej.test_util.TestImgGeneration;
+import net.imagej.testutil.TestImgGeneration;
 import net.imglib2.img.Img;
 import net.imglib2.img.array.ArrayImg;
 import net.imglib2.img.array.ArrayImgs;
