+++ pom.xml
@@ -46,9 +46,9 @@
 
 	<modules>
 		<module>imagej/imagej-ops2</module>
-		<module>imagej/imagej-test-util</module>
+		<module>imagej/imagej-testutil</module>
 		<module>scijava/scijava-ops</module>
-		<module>scijava/scijava-test-util</module>
+		<module>scijava/scijava-testutil</module>
 		<module>scijava/scijava-types</module>
 	</modules>
