+++ scijava/scijava-testutil/src/main/java/org/scijava/testutil/AssertIterations.java
@@ -1,4 +1,4 @@
-package org.scijava.test_util;
+package org.scijava.testutil;
 
 import java.util.Iterator;
