+++ scijava/scijava-testutil/src/test/java/org/scijava/testutil/AppTest.java
@@ -1,4 +1,4 @@
-package org.scijava.test_util;
+package org.scijava.testutil;
 
 import static org.junit.Assert.assertTrue;
